Create commons module, start refactoring into it, create kafka-producer (kafka-producer will take advantage of the refactoring)
