+++ commons/src/main/java/com/expedia/www/haystack/pipes/commons/Configuration.java
@@ -1,0 +1,33 @@
+package com.expedia.www.haystack.pipes.commons;
+
+import org.cfg4j.provider.ConfigurationProvider;
+import org.cfg4j.provider.ConfigurationProviderBuilder;
+import org.cfg4j.source.ConfigurationSource;
+import org.cfg4j.source.classpath.ClasspathConfigurationSource;
+import org.cfg4j.source.compose.MergeConfigurationSource;
+import org.cfg4j.source.system.EnvironmentVariablesConfigurationSource;
+
+import java.nio.file.Paths;
+import java.util.Collections;
+
+public class Configuration {
+    public static final String HAYSTACK_GRAPHITE_CONFIG_PREFIX = "haystack.graphite";
+
+    public static ConfigurationProvider createMergeConfigurationProvider() {
+        final MergeConfigurationSource configurationSource = new MergeConfigurationSource(
+                createClasspathConfigurationSource(), createEnvironmentConfigurationSource()
+        );
+        final ConfigurationProviderBuilder configurationProviderBuilder = new ConfigurationProviderBuilder();
+        return configurationProviderBuilder.withConfigurationSource(configurationSource).build();
+    }
+
+    private static ConfigurationSource createClasspathConfigurationSource() {
+        return new ClasspathConfigurationSource(() -> Collections.singletonList(Paths.get("base.yaml")));
+    }
+
+    private static ConfigurationSource createEnvironmentConfigurationSource() {
+        final EnvironmentVariablesConfigurationSource environmentVariablesConfigurationSource =
+                new EnvironmentVariablesConfigurationSource();
+        return new ChangeEnvVarsToLowerCaseConfigurationSource("HAYSTACK", environmentVariablesConfigurationSource);
+    }
+}
