+++ json-transformer/pom.xml
@@ -17,6 +17,7 @@
         <cfg4j-core-version>4.4.1</cfg4j-core-version>
         <commons-text-version>1.1</commons-text-version>
         <haystack-metrics-version>0.1</haystack-metrics-version>
+        <haystack-pipes-commons-version>1.0-SNAPSHOT</haystack-pipes-commons-version>
         <jacoco-maven-plugin-version>0.7.9</jacoco-maven-plugin-version>
         <java-version>1.8</java-version>
         <junit-version>4.12</junit-version>
@@ -42,6 +43,11 @@
     </parent>
 
     <dependencies>
+        <dependency>
+            <groupId>com.expedia.www</groupId>
+            <artifactId>haystack-pipes-commons</artifactId>
+            <version>${haystack-pipes-commons-version}</version>
+        </dependency>
         <dependency>
             <groupId>org.apache.kafka</groupId>
             <artifactId>kafka_2.11</artifactId>
